move click fns to userland
