+++ src/lib.rs
@@ -1,9 +1,6 @@
 #![allow(clippy::type_complexity)]
 
 use std::{collections::VecDeque, path::PathBuf, time::Duration};
-#[path = "utils.rs"]
-pub mod utils;
-pub use utils::*;
 
 use bevy::{
     asset::HandleId,
@@ -1506,6 +1503,19 @@ fn draw_pixel(
     buffer[offset + 1] = (current >> 8) as u8;
     buffer[offset] = (current >> 16) as u8;
     buffer[offset + 3] = (current >> 24) as u8;
+}
+
+#[cfg(target_arch = "wasm32")]
+pub fn get_timestamp() -> f64 {
+    js_sys::Date::now()
+}
+
+#[cfg(not(target_arch = "wasm32"))]
+pub fn get_timestamp() -> f64 {
+    use std::time::SystemTime;
+    use std::time::UNIX_EPOCH;
+    let duration = SystemTime::now().duration_since(UNIX_EPOCH).unwrap();
+    duration.as_millis() as f64
 }
 
 #[cfg(test)]
